requirements.txt
fastapi~=0.61.2
uvicorn[standart]~=0.12.2
pytest-asyncio
pandas==1.1.5
orjson==3.4.6
requests==2.25.1
nest-asyncio==1.4.3
